requirements.txt
shapely
python-craigslist